[STABLE] Hardware (pin) toggle buttons on control page display pin status
